If you don't want to use a resource at least make sure to have a helper
Most of the times is better to go with restful resources, that will make
sure you are following th convention for controller organization that
rails tries to enforce.

I even try to avoid adding extra actions that are not part of the 7 default
actions, by doing that I force myself to properly think about the resources
and that helps me avoid gigantic controllers with many different actions.

After doing this exercise of trying to avoid extra actions and also
trying think about your controllers as resources, if you still think
it is better to go with a non-restful controller at least provide a
helper for that url (using the option as: ) so you don't have to
hardcode urls/paths in your code.