Design: Write your msgs as a human being not a database
Rails will generate the messages in a not so friendly way beucase there
is not way for the framework to do it. It's your job to make adjust all
the copy to be more human, remember this talk:
https://www.youtube.com/watch?v=sz_LgBAGYyo. Copy is one of the main
areas to impact the experience so try to no sound like a robot.