Be consistent, or we handle json requests or we don't
Rails comes with the native support to handle json requests if you use
scaffold. I love this feature and try to make sure it works even when I
customize the generated code but that's something you have to do
manually. And I don't think we gonna worry about providing an API for
this project any time soon, and since most of the controllers are not
handling json requests/responses properly already I'm just gonna remove
the respond to block.

In theory we are making our controller not rest
anymore because the REST concept assumes that a resource can be accessible using many
formats instead of just plain html, for example, another software we
probably not be able to handle html like a browser does. So that's a
trade-off and for now I carry more about serving html properly than
supporting different types of accesses.