Adding reports association to project and other tweaks
I prefer to use has_attached_file with styles using the size with # at
the end, that is a resize and crop instead of a hard resize that doesn't
respect the ratio (like when you use !). I'm bringing back the
validation of the owner, turning the owner_email method in a one liner
and also putting all the code associated to screenshots in the same
area. I also remove the paginate 10 because 30 projects per page is
perfectly fine from performance stand point.