Bad security breach in ReleasesController
Every time you write a controller you should pay attention to two
things: Who should be able to read the data and who should be to change
the data. In this guest visitors should be able to read (show action)
but only project owners should be to change (all other actions). That's
an authorization problem.

There's many ways to handle authorization but if the rules are not too
complex like many different types of user I tend to prefer just to scope
the queries to the proper user like I'm doing here.