Changing a release from open to closed is just an update
There's no reason to creates methods on the model and then an extra
action if you think about the resources. Closed or not for is just a
change of an atribute on the release resource. In cases like that, when
the change is simple I prefer to just go with the normal update action
of the resource.

In other situations where the change is more complex,
for example a CMS where multiple users could set the status of a blog post
to draft, reviewing, reviewed and published I would probably go with a
controller for the status, like PostStatus but when the change is simple
like in our case I think even an extra action is unecessary
complication.

In the previous commit I also introduced a bug with the filters, so this
commit is a refactoring of close the release and also a bug fix