Extracting report status to a different controller
The necessity to extra actions to the controller is a good indicator
that your controller might be dealing with more than one resource which
tends to make your controllers bloated. For example, dealing with the
status has some special rules that doesn't apply to all the other
actions we have in the reports controllers, in this case I prefer to
extract a resource (ReportStatus) to deal just with the status.

The original code in the reports_controller had a security bug that
allowed other people that are not project owners to accept or deny the
reports, cheating the leaderboard. I didn't fix the bug on purpose.