Refactoring leaderboard to a different class
One simple principle (not law) you should keep in mind when writing code is
Single Responsibility. Your code will be (in theory) easier to
understand if does only one thing. A method should do only one thing,
one class should take care of just one object, one controller should
handle just one main resource, so on and so forth.

To me, mixing the User model with the idea of the Leaderboard is not
really necessary and adds extra logic that is not exclusive to the
user.With that said I decided to extract the Leaderboard to it's own
model.

Models, by default on a Rails app, inherit from
ActiveRecord::Base and that gives the database integration we are used
to but sometimes you just need a class/object instead of a database
table. That's the perfect case for the Leaderboard class.

I detailed the implementation of the Leaderboard class with comments but
we basically have just two methods. One if to fetch the current
leaderboard from the database (combining Users and Accepted Reports) and
the other is just to find the position of one specific user in the
Leaderboard.

The difference of my implemenation (using the fetch method) is mainly
based on the fact that I'm relying on the database to count and give me
back a result set that represents the leaderboard instead of getting an
array of things creating everything with ruby. Database is more
efficient for this type of calculation than loops in Ruby. So the rule
of thumb is to delegate most of the data manipulation to the database as
most as possible and that tends to lead us to easier to read code and
also more efficient. If you don't understand the sql to create the
result set you should research about JOINS, Count and Select.

HINT: If you call Leaderboard.fetch from the console you'll see the generated SQL
wich will give you a better hint for you research about sql.